Add files via upload
This script takes two .dat files outputted by turboSETI and finds the matching events between the two files, creating waterfall plots for all events and saving each plot as a .png file. Two arguments are required to run this script: the first should be the path to the directory containing two .dat files, and the second should be the path to the directory containing two corresponding .h5 files.